simplifying void Main() - dropping Program.cs and moving it to Startup.cs
