--- samples/core/GetStarted/AspNetCore/AspNetCore.ExistingDb/Program.cs
@@ -1,27 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Configuration.UserSecrets;
-
-[assembly: UserSecretsId("aspnet-AspNetCore.ExistingDb-20161230022416")]
-
-namespace EFGetStarted.AspNetCore.ExistingDb
-{
-    public class Program
-    {
-        public static void Main(string[] args)
-        {
-            var host = new WebHostBuilder()
-                .UseKestrel()
-                .UseContentRoot(Directory.GetCurrentDirectory())
-                .UseIISIntegration()
-                .UseStartup<Startup>()
-                .Build();
-
-            host.Run();
-        }
-    }
-}
